updated checking functions

+++ vid/vid_start_end.py
@@ -8,6 +8,7 @@ import re
 import subprocess
 import pdb
 import cPickle as pickle
+import numpy as np
 
 def get_len(filename):
     
@@ -51,47 +52,47 @@ def check_disconnect(cur_time, vid_len, start):
 def main():
    root_folder = "D:\\NancyStudyData\\ecog\\raw\\"
    sbj_ids = ['a86a4375', 'be66b17c', 'cb46fd46', 'da3971ee', 'fcb01f7a']
-   days = ['3','4','5','6','7','8','9','10']
-   sbj_ids = ['a86a4375']
-   days = ['2']
+   sbj_ids = ['fcb01f7a']
+   days = ['1','2','3','4','5','6','7','8','9','10', '11','12','13', '16']
    save_folder = "C:\\Users\\wangnxr\\Documents\\rao_lab\\video_analysis\\vid_real_time\\"
 
    for sbj_id in sbj_ids:
        for day in days:
-            result = []
-            start_time, end_time, start, end = get_disconnected_times("C:\\Users\\wangnxr\\Documents" + \
+            disconnect_file= "C:\\Users\\wangnxr\\Documents" + \
                                "\\rao_lab\\video_analysis\\disconnect_times\\" \
-                               + sbj_id + "_" + str(day) + ".txt")
-           
-            vid_count = 0
-            cur_time = start_time
-            vid_name = root_folder + sbj_id + "\\" + sbj_id + "_" + day + \
-                            "\\" + sbj_id + "_" + day + "_" + str(vid_count).zfill(4) + ".avi"
-            with open(save_folder + sbj_id + "_" + day + ".csv", "wb") as csvfile:
-               timewriter = csv.writer(csvfile)
-               pos = check_disconnect(cur_time, timedelta(seconds = 1), start)
-               for t in pos:
-                   cur_time += end[t] - start[t] - timedelta(seconds = 0.5)
-               video_start_times = []
-               video_end_times = []
-               while os.path.exists(vid_name):
-                   timewriter.writerow([vid_name, cur_time.year, cur_time.month,
-                                        cur_time.day, cur_time.hour, cur_time.minute, cur_time.second])
-                   video_start_times.append(cur_time)
-                   vid_len = timedelta(seconds = get_len(root_folder + sbj_id + "\\" + sbj_id + "_" + day +
-                                    "\\" + sbj_id + "_" + day + "_" + str(vid_count).zfill(4) + ".avi"))
-                   
-                   pos = check_disconnect(cur_time, vid_len, start)
-                   cur_time += vid_len
-                   video_end_times.append(cur_time)
+                               + sbj_id + "_" + str(day) + ".txt"
+            if os.path.isfile(disconnect_file):
+                print disconnect_file
+                start_time, end_time, start, end = get_disconnected_times(disconnect_file)
+                vid_count = 0
+                cur_time = start_time
+                vid_name = root_folder + sbj_id + "\\" + sbj_id + "_" + day + \
+                                "\\" + sbj_id + "_" + day + "_" + str(vid_count).zfill(4) + ".avi"
+                with open(save_folder + sbj_id + "_" + day + ".csv", "wb") as csvfile:
+                   timewriter = csv.writer(csvfile)
+                   pos = check_disconnect(cur_time, timedelta(seconds = 1), start)
                    for t in pos:
                        cur_time += end[t] - start[t] - timedelta(seconds = 0.5)
+                   video_start_times = []
+                   video_end_times = []
+                   while os.path.exists(vid_name):
+                       timewriter.writerow([vid_name, cur_time.year, cur_time.month,
+                                            cur_time.day, cur_time.hour, cur_time.minute, cur_time.second])
+                       video_start_times.append(cur_time)
+                       vid_len = timedelta(seconds = get_len(root_folder + sbj_id + "\\" + sbj_id + "_" + day +
+                                        "\\" + sbj_id + "_" + day + "_" + str(vid_count).zfill(4) + ".avi"))
 
-                   vid_count += 1
-                   vid_name = root_folder + sbj_id + "\\" + sbj_id + "_" + day + \
-                              "\\" + sbj_id + "_" + day + "_" + str(vid_count).zfill(4) + ".avi"
-               result = {"start": np.array(video_start_times), "end": np.array(video_end_times)}
-            pickle.dump(result, open(save_folder + sbj_id + "_" + day + ".p", "wb"))
+                       pos = check_disconnect(cur_time, vid_len, start)
+                       cur_time += vid_len
+                       video_end_times.append(cur_time)
+                       for t in pos:
+                           cur_time += end[t] - start[t] - timedelta(seconds = 0.5)
+
+                       vid_count += 1
+                       vid_name = root_folder + sbj_id + "\\" + sbj_id + "_" + day + \
+                                  "\\" + sbj_id + "_" + day + "_" + str(vid_count).zfill(4) + ".avi"
+                   result = {"start": np.array(video_start_times), "end": np.array(video_end_times)}
+                pickle.dump(result, open(save_folder + sbj_id + "_" + day + ".p", "wb"))
             
 if __name__ == "__main__":
    main()
